+++ test/test.js
@@ -4,6 +4,7 @@ var http = require('http')
 var path = require('path')
 var request = require('supertest')
 var serveStatic = require('..')
+var uaParser = require('ua-parser');
 
 var fixtures = path.join(__dirname, '/fixtures')
 var relative = path.relative(process.cwd(), fixtures)
@@ -778,10 +779,15 @@ describe('serveStatic()', function () {
   })
 })
 
+function differentialServeMiddleware (req, res, next, opts){
+  opts.root = opts.root + parseUA(req.headers['user-agent'])
+  console.log(opts);
+}
+
 function createServer (dir, opts, fn) {
   dir = dir || fixtures
 
-  var _serve = serveStatic(dir, opts)
+  var _serve = serveStatic(dir, opts, [differentialServeMiddleware])
 
   return http.createServer(function (req, res) {
     fn && fn(req, res)
@@ -796,4 +802,24 @@ function shouldNotHaveHeader (header) {
   return function (res) {
     assert.ok(!(header.toLowerCase() in res.headers), 'should not have header ' + header)
   }
+}
+
+
+/**
+ * Parse a user-agent and return the base-directory
+ * @private
+ */
+function parseUA(ua){
+  var uaParsed = uaParser.parseUA(ua)
+
+  const browser = uaParsed.family
+  const majorVersion = uaParsed.major
+
+  const supportsES2015 = (browser === 'Chrome' && majorVersion >= 49) ||
+    (browser === 'Safari' && majorVersion >= 10) ||
+    (browser === 'Edge' && majorVersion >= 14) ||
+    (browser === 'Firefox' && majorVersion >= 51)
+
+  return supportsES2015 ? '/es2015' : '/es5'
+
 }
